move binning and  io to modules
