+++ sed/binning/__init__.py
@@ -1,0 +1,4 @@
+from .binning import bin_dataframe
+from .binning import bin_partition
+
+__all__ = ["bin_dataframe", "bin_partition"]
