+++ sed/binning/binning.py
@@ -1,5 +1,3 @@
-# Note: some of the functions presented here were
-# inspired by https://github.com/mpes-kit/mpes
 from functools import reduce
 from typing import List
 from typing import Sequence
@@ -7,7 +5,6 @@ from typing import Tuple
 from typing import Union
 
 import dask.dataframe
-import numba
 import numpy as np
 import pandas as pd
 import psutil
@@ -15,149 +12,11 @@ import xarray as xr
 from threadpoolctl import threadpool_limits
 from tqdm.auto import tqdm
 
+from .numba_bin import numba_histogramdd
+from .utils import _arraysum
+from .utils import _simplify_binning_arguments
+
 N_CPU = psutil.cpu_count()
-
-
-def _arraysum(array_a, array_b):
-    """Calculate the sum of two arrays."""
-    return array_a + array_b
-
-
-def _simplify_binning_arguments(
-    bins: Union[
-        int,
-        dict,
-        tuple,
-        List[int],
-        List[np.ndarray],
-        List[tuple],
-    ] = 100,
-    axes: Union[str, Sequence[str]] = None,
-    ranges: Sequence[Tuple[float, float]] = None,
-) -> tuple:
-    """ Convert the flexible input for defining bins into a 
-    simple "axes" "bins" "ranges" tuple.
-    
-    This allows to mimic the input used in numpy.histogramdd flexibility into the 
-    binning functions defined here.
-
-    Args:
-        bins: Definition of the bins. Can  be any of the following cases:
-            - an integer describing the number of bins in on all dimensions
-            - a tuple of 3 numbers describing start, end and step of the binning range
-            - a np.arrays defining the binning edges
-            - a list (NOT a tuple) of any of the above (int, tuple or np.ndarray)
-            - a dictionary made of the axes as keys and any of the above as values.
-            This takes priority over the axes and range arguments. Defaults to 100
-        axes: The names of the axes (columns) on which to calculate the histogram.
-            The order will be the order of the dimensions in the resulting array.
-            Defaults to None
-        ranges: list of tuples containing the start and end point of the binning range.
-            Defaults to None
-
-    Returns:
-        tuple containing axes, bins and ranges.
-    """
-    if isinstance(axes,str):
-        axes = [axes]
-    # if bins is a dictionary: unravel to axes and bins
-
-    if isinstance(bins, dict):  
-        axes = []
-        bins_ = []
-        for k, v in bins.items():
-            axes.append(k)
-            bins_.append(v)
-        bins = bins_
-    elif isinstance(bins, (int, np.ndarray)):
-        bins = [bins] * len(axes)
-    elif isinstance(bins, tuple):
-        if len(bins) == 3:
-            bins = [bins]
-        else:
-            raise ValueError(
-                "Bins defined as tuples should only be used to define start "
-                / "stop and step of the bins. i.e. should always have lenght 3.",
-            )
-    if not isinstance(bins, list):
-        raise TypeError(f"Cannot interpret bins of type {type(bins)}")
-    if axes is None:
-        raise AttributeError("Must define on which axes to bin")
-    if not all(isinstance(x, type(bins[0])) for x in bins):
-        raise TypeError("All elements in " "bins must be of the same type")
-
-    bin = bins[0]
-    if isinstance(bin, tuple):
-        ranges = []
-        bins_ = []
-        for tpl in bins:
-            ranges.append([tpl[0], tpl[1]])
-            bins_.append(tpl[2])
-        bins = bins_
-    elif not isinstance(bin, (int, np.ndarray)):
-        raise TypeError(f"Could not interpret bins of type {type(bin)}")
-
-    if ranges is not None:
-        if (len(axes) == len(bins) == 1) and len(ranges) == 2:
-            pass
-        elif not (len(axes) == len(bins) == len(ranges)):
-            raise AttributeError(
-                "axes and range and bins must have the same number of elements",
-            )
-    elif isinstance(bin, int):
-        raise AttributeError(
-            "Must provide a range if bins is an integer or list of integers",
-        )
-    elif len(axes) != len(bins):
-        raise AttributeError(
-            "axes and bins must have the same number of elements",
-        )
-
-    return bins, axes, ranges
-
-
-def bin_edges_to_bin_centers(bin_edges: np.array) -> np.array:
-    """Converts a list of bin edges into corresponding bin centers
-
-    Args:
-        bin_edges: 1d array of bin edges
-
-    Returns:
-        bin_centers: 1d array of bin centers
-    """
-
-    bin_centers = (bin_edges[1:] + bin_edges[:-1]) / 2
-
-    return bin_centers
-
-
-def bin_centers_to_bin_edges(bin_centers: np.ndarray) -> np.ndarray:
-    """Converts a list of bin centers into corresponding bin edges
-
-    Args:
-        bin_centers: 1d array of bin centers
-
-    Returns:
-        bin_edges: 1d array of bin edges
-    """
-    bin_edges = (bin_centers[1:] + bin_centers[:-1]) / 2
-
-    bin_edges = np.insert(
-        bin_edges,
-        0,
-        bin_centers[0] - (bin_centers[1] - bin_centers[0]) / 2,
-    )
-    bin_edges = np.append(
-        bin_edges,
-        bin_centers[len(bin_centers) - 1]
-        + (
-            bin_centers[len(bin_centers) - 1]
-            - bin_centers[len(bin_centers) - 2]
-        )
-        / 2,
-    )
-
-    return bin_edges
 
 
 def bin_partition(
@@ -498,241 +357,3 @@ def apply_jitter_on_column(
         # transformations and jitter sizes that don't match the original bin
         # sizes
         df[col] += amp * np.random.standard_normal(size=colsize)
-
-
-@numba.jit(nogil=True, nopython=True)
-def _hist_from_bin_range(
-    sample: np.array,
-    bins: Sequence,
-    ranges: np.array,
-) -> np.array:
-    """
-    N dimensional binning function, pre-compiled by Numba for performance.
-    Behaves much like numpy.histogramdd, but calculates and returns unsigned 32
-    bit integers.
-
-    Args:
-        sample: The data to be histogrammed with shape N,D.
-        bins: the number of bins for each dimension D.
-        range: A sequence of length D, each an optional (lower,
-            upper) tuple giving the outer bin edges to be used if the edges are
-            not given explicitly in bins.
-
-    Raises:
-        ValueError: In case of dimension mismatch.
-
-    Returns:
-        The computed histogram.
-    """
-    ndims = len(bins)
-    if sample.shape[1] != ndims:
-        raise ValueError(
-            "The dimension of bins is not equal to the dimension of the sample x",
-        )
-
-    H = np.zeros(bins, np.uint32)
-    Hflat = H.ravel()
-    delta = np.zeros(ndims, np.float64)
-    strides = np.zeros(ndims, np.int64)
-
-    for i in range(ndims):
-        delta[i] = 1 / ((ranges[i, 1] - ranges[i, 0]) / bins[i])
-        strides[i] = H.strides[i] // H.itemsize
-
-    for t in range(sample.shape[0]):
-        is_inside = True
-        flatidx = 0
-        for i in range(ndims):
-            j = (sample[t, i] - ranges[i, 0]) * delta[i]
-            is_inside = is_inside and (0 <= j < bins[i])
-            flatidx += int(j) * strides[i]
-            # don't check all axes if you already know you're out of the range
-            if not is_inside:
-                break
-        if is_inside:
-            Hflat[flatidx] += int(is_inside)
-
-    return H
-
-
-@numba.jit(nogil=True, parallel=False, nopython=True)
-def binsearch(bins: np.ndarray, val: float) -> int:
-    """Bisection index search function.
-
-    Finds the index of the bin with the highest value below val, i.e. the left edge.
-    returns -1 when the value is outside the bin range.
-
-    Args:
-        bins: the array on which
-        val: value to search for
-
-    Returns:
-        int: index of the bin array, returns -1 when value is outside the bins range
-    """
-    low, high = 0, len(bins) - 1
-    mid = high // 2
-    if val == bins[high]:
-        return high - 1
-    if (val < bins[low]) | (val > bins[high]):
-        return -1
-
-    while True:
-        if val < bins[mid]:
-            high = mid
-        elif val < bins[mid + 1]:
-            return mid
-        else:
-            low = mid
-        mid = (low + high) // 2
-
-
-@numba.jit(nopython=True, nogil=True, parallel=False)
-def _hist_from_bins(
-    sample: np.ndarray,
-    bins: Sequence[np.ndarray],
-    shape: Tuple,
-) -> np.ndarray:
-    """Numba powered binning method, similar to np.histogramdd.
-
-    Computes the histogram on pre-defined bins.
-
-    Args:
-        sample : the array of shape (N,D) on which to compute the histogram
-        bins : array of shape (N,D) defining the D bins on which to compute
-            the histogram, i.e. the desired output axes.
-        shape: shape of the resulting array. Workaround for the fact numba
-            does not allow to create tuples.
-    Returns:
-        hist : the computed n-dimensional histogram
-    """
-    ndims = len(bins)
-    if sample.shape[1] != ndims:
-        raise ValueError(
-            "The dimension of bins is not equal to the dimension of the sample x",
-        )
-    H = np.zeros(shape, np.uint32)
-    Hflat = H.ravel()
-
-    strides = np.zeros(ndims, np.int64)
-
-    for i in range(ndims):
-        strides[i] = H.strides[i] // H.itemsize
-    for t in range(sample.shape[0]):
-        is_inside = True
-        flatidx = 0
-        for i in range(ndims):
-            j = binsearch(bins[i], sample[t, i])
-            # binsearch returns -1 when the value is outside the bin range
-            is_inside = is_inside and (j >= 0)
-            flatidx += int(j) * strides[i]
-            # don't check all axes if you already know you're out of the range
-            if not is_inside:
-                break
-        if is_inside:
-            Hflat[flatidx] += int(is_inside)
-
-    return H
-
-
-def numba_histogramdd(
-    sample: np.array,
-    bins: Union[Sequence[Union[int, np.ndarray]], np.ndarray],
-    ranges: Sequence = None,
-) -> Tuple[np.ndarray, np.ndarray]:
-    """Multidimensional histogramming function, powered by Numba.
-
-    Behaves in total much like numpy.histogramdd. Returns uint32 arrays.
-    This was chosen because it has a significant performance improvement over
-    uint64 for large binning volumes. Be aware that this can cause overflows
-    for very large sample sets exceeding 3E9 counts in a single bin. This
-    should never happen in a realistic photoemission experiment with useful bin
-    sizes.
-
-    Args:
-        sample: The data to be histogrammed with shape N,D
-        bins: the number of bins for each dimension D, or a sequence of bins
-        on which to calculate the histogram.
-        range: The range to use for binning when bins is a list of integers.
-
-    Raises:
-        ValueError: In case of dimension mismatch.
-        NotImplementedError: When attempting binning in too high number of
-        dimensions (>4)
-        RuntimeError: Internal shape error after binning
-
-    Returns:
-        2-element tuple returned only when returnEdges is True. Otherwise
-        only hist is returned.
-
-        - **hist**: The computed histogram
-        - **edges**: A list of D arrays describing the bin edges for
-            each dimension.
-    """
-
-    try:
-        # Sample is an ND-array.
-        N, D = sample.shape
-    except (AttributeError, ValueError):
-        # Sample is a sequence of 1D arrays.
-        sample = np.atleast_2d(sample).T
-        N, D = sample.shape
-
-    if not isinstance(bins, (tuple, list)):
-        bins = D * [bins]
-    Db = len(bins)
-    if isinstance(bins[0], (int, np.int_)):
-        method = "int"
-    elif isinstance(bins[0], np.ndarray):
-        method = "array"
-    else:
-        raise AttributeError(
-            f"bins must be int, np.ndarray or a sequence of the two. "
-            f"Found {type(bins[0])} instead",
-        )
-
-    if Db != D:  # check number of dimensions
-        raise ValueError(
-            "The dimension of bins must be equal to the dimension of the sample x.",
-        )
-
-    if method == "array":
-        hist = _hist_from_bins(
-            sample,
-            tuple(bins),
-            tuple(b.size - 1 for b in bins),
-        )
-        return hist, bins
-
-    elif method == "int":
-        # normalize the range argument
-        if ranges is None:
-            raise ValueError(
-                "must define a value for ranges when bins is"
-                " the number of bins",
-            )
-        #     ranges = (None,) * D
-        if D == 1 and isinstance(ranges[0], (int, float)):
-            ranges = (ranges,)
-        elif len(ranges) != D:
-            raise ValueError(
-                "range argument must have one entry per dimension",
-            )
-
-        ranges = np.asarray(ranges)
-        bins = tuple(bins)
-
-        # Create edge arrays
-        edges = D * [None]
-        nbin = np.empty(D, int)
-
-        for i in range(D):
-            edges[i] = np.linspace(*ranges[i, :], bins[i] + 1)
-
-            nbin[i] = len(edges[i]) + 1  # includes an outlier on each end
-
-        hist = _hist_from_bin_range(sample, bins, ranges)
-
-        if (hist.shape != nbin - 2).any():
-            raise RuntimeError("Internal Shape Error")
-
-        return hist, edges
