+++ sed/core/__init__.py
@@ -1,0 +1,3 @@
+from .processor import SedProcessor
+
+__all__ = ["SedProcessor"]
