+++ sed/io/__init__.py
@@ -1,0 +1,11 @@
+from .hdf5 import load_h5
+from .hdf5 import to_h5
+from .tiff import load_tiff
+from .tiff import to_tiff
+
+__all__ = [
+    "load_h5",
+    "to_h5",
+    "load_tiff",
+    "to_tiff",
+]
